EES-3101 Change how DropAndCreateRelease deletes Subject rows

+++ src/GovUk.Education.ExploreEducationStatistics.Data.Model/Migrations/20220114144422_Routine_DropAndCreateRelease.sql
@@ -1,33 +1,33 @@
-CREATE OR ALTER PROCEDURE DropAndCreateRelease @Release dbo.ReleaseType READONLY, @LinkedSubjects varchar(4000), @LinkedFootnotes varchar(4000)
+CREATE OR ALTER PROCEDURE DropAndCreateRelease @Release dbo.ReleaseType READONLY
 AS
 BEGIN
-    DECLARE @LinkedSubjectsList IdListGuidType,
-        @LinkedFootnotesList IdListGuidType;
+    DECLARE @ReleaseId AS UNIQUEIDENTIFIER;
+    SET @ReleaseId = (SELECT R.Id FROM @Release R);
 
-    IF LEN(ISNULL(@LinkedSubjects, '')) > 0
-        INSERT @LinkedSubjectsList SELECT value FROM STRING_SPLIT(@LinkedSubjects,',');
+    IF @ReleaseId IN (SELECT Id FROM dbo.Release R WHERE R.Id = @ReleaseId)
+        BEGIN
+            DECLARE @FootnoteIds TABLE (Id UNIQUEIDENTIFIER PRIMARY KEY NOT NULL);
+            INSERT INTO @FootnoteIds (Id) SELECT RF.FootnoteId FROM dbo.ReleaseFootnote RF WHERE ReleaseId = @ReleaseId
 
-    IF LEN(ISNULL(@LinkedFootnotes, '')) > 0
-        INSERT @LinkedFootnotesList SELECT value FROM STRING_SPLIT(@LinkedFootnotes,',');
+            DELETE dbo.FilterFootnote FROM dbo.FilterFootnote FF JOIN @FootnoteIds F ON F.Id = FF.FootnoteId;
+            DELETE dbo.FilterGroupFootnote FROM dbo.FilterGroupFootnote FGF JOIN @FootnoteIds F ON F.Id = FGF.FootnoteId;
+            DELETE dbo.FilterItemFootnote FROM dbo.FilterItemFootnote FIF JOIN @FootnoteIds F ON F.Id = FIF.FootnoteId;
+            DELETE dbo.IndicatorFootnote FROM dbo.IndicatorFootnote INF JOIN @FootnoteIds F ON F.Id = INF.FootnoteId;
+            DELETE dbo.SubjectFootnote FROM dbo.SubjectFootnote SF JOIN @FootnoteIds F ON F.Id = SF.FootnoteId;
 
-    DECLARE @FootnoteIds TABLE (Id UNIQUEIDENTIFIER PRIMARY KEY NOT NULL);
-    INSERT INTO @FootnoteIds (Id) SELECT RF.FootnoteId FROM dbo.ReleaseFootnote RF WHERE ReleaseId = (SELECT R.Id FROM @Release R);
+            DELETE FROM dbo.ReleaseFootnote WHERE dbo.ReleaseFootnote.ReleaseId = @ReleaseId;
 
-    DELETE FROM dbo.SubjectFootnote WHERE FootnoteId IN (SELECT F.Id FROM @FootnoteIds F);
-    DELETE FROM dbo.FilterFootnote WHERE FootnoteId IN (SELECT F.Id FROM @FootnoteIds F);
-    DELETE FROM dbo.FilterGroupFootnote WHERE FootnoteId IN (SELECT F.Id FROM @FootnoteIds F);
-    DELETE FROM dbo.FilterItemFootnote WHERE FootnoteId IN (SELECT F.Id FROM @FootnoteIds F);
-    DELETE FROM dbo.IndicatorFootnote WHERE FootnoteId IN (SELECT F.Id FROM @FootnoteIds F);
-    DELETE FROM dbo.SubjectFootnote WHERE FootnoteId IN (SELECT F.Id FROM @FootnoteIds F);
+            DELETE FROM dbo.Footnote WHERE Id IN (SELECT F.Id FROM @FootnoteIds F);
 
-    DELETE FROM dbo.ReleaseFootnote WHERE dbo.ReleaseFootnote.ReleaseId IN (SELECT R.Id FROM @Release R);
+            DECLARE @SubjectIds TABLE (Id UNIQUEIDENTIFIER PRIMARY KEY NOT NULL);
+            INSERT INTO @SubjectIds (Id) SELECT RS.SubjectId FROM dbo.ReleaseSubject RS WHERE ReleaseId = @ReleaseId
 
-    DELETE FROM dbo.Footnote WHERE Id IN (SELECT F.Id FROM @FootnoteIds F);
+            DELETE FROM dbo.ReleaseSubject WHERE dbo.ReleaseSubject.ReleaseId = @ReleaseId
+            DELETE FROM dbo.Subject WHERE dbo.Subject.Id IN (SELECT S.Id FROM @SubjectIds S)
+                                      AND dbo.Subject.Id NOT IN (Select RS.SubjectId FROM dbo.ReleaseSubject RS);
 
-    DELETE FROM dbo.ReleaseSubject WHERE dbo.ReleaseSubject.ReleaseId IN (SELECT R.Id FROM @Release R);
-    DELETE FROM dbo.Release WHERE dbo.Release.Id IN (SELECT R.Id FROM @Release R);
-
-    DELETE FROM dbo.Subject WHERE dbo.Subject.Id NOT IN (SELECT id FROM @LinkedSubjectsList) AND NOT EXISTS (SELECT * FROM dbo.ReleaseSubject WHERE dbo.ReleaseSubject.SubjectId = dbo.Subject.Id) AND SoftDeleted != 1;
+            DELETE FROM dbo.Release WHERE dbo.Release.Id = @ReleaseId;
+        END
 
     INSERT INTO dbo.Release (Id, TimeIdentifier, Slug, Year, PublicationId, PreviousVersionId)
     SELECT Id, TimeIdentifier, Slug, Year, PublicationId, PreviousVersionId FROM @Release;
